refactor: move custom components to component directory
