+++ src/App.test.js
@@ -1,7 +1,7 @@
 import { render, screen } from "@testing-library/react";
 import { BrowserRouter } from "react-router-dom";
 import App from "./App";
-import SearchResultPage from "./SearchResultPage";
+import SearchResultPage from "./Components/SearchResultPage";
 
 test("renders landing page heading", () => {
   render(<App />);
